Agregar nuevas vistas admin, mejorar vistas cliente

+++ src/api/axiosConfig.js
@@ -29,9 +29,16 @@ api.interceptors.response.use(
   response => response,
   error => {
     if (error.response && (error.response.status === 401 || error.response.status === 403)) {
-      console.warn("Sesión expirada o no autorizada. Redirigiendo al login...");
-      localStorage.clear();
-      globalThis.location.href = '/login';
+      
+      // Verificamos de qué endpoint vino el error
+      const isLoginRequest = error.config && error.config.url && error.config.url.includes('/auth/login');
+      
+      // Si NO es un intento de login, entonces sí limpiamos y recargamos
+      if (!isLoginRequest) {
+        console.warn("Sesión expirada o no autorizada. Redirigiendo al login...");
+        localStorage.clear();
+        globalThis.location.href = '/login';
+      }
     }
     return Promise.reject(error);
   }
